Moved bill class modal to its own file
